# Conflicts:
#	src/movie/Movie.java
#	src/movie/MovieCollection.java

+++ src/commands/Remove_if_lowe.java
@@ -10,7 +10,7 @@ public class Remove_if_lowe implements ICommand {
     @Override
     public void Do(String key, String parameter2) {
         MovieCollection movieCollection = new MovieCollection();
-        for (int keyToRemove = 0; keyToRemove < Integer.parseInt(key); keyToRemove++) {
+        for (long keyToRemove = 0; keyToRemove < Integer.parseInt(key); keyToRemove++) {
             movieCollection.removeMovie(keyToRemove);
         }
     }
