+++ src/commands/update_id.java
@@ -1,6 +1,7 @@
 package commands;
 
 import data.MovieCollection;
+import movie.Location;
 
 public class Update_id implements ICommand {
     public Update_id(){
@@ -10,7 +11,7 @@ public class Update_id implements ICommand {
     @Override
     public void Do(String parameter1, String parameter2){
         MovieCollection movieCollection = new MovieCollection();
-        Integer key = Integer.parseInt (parameter1.trim ());
+        Long key = Long.parseLong (parameter1.trim ());
         movieCollection.replaceMovie(key, parameter2);
     }
 }
